Movendo o UserManage para um arquivo proprio #6
close #6
